+++ lang/parse/parse.go
@@ -644,6 +644,22 @@ func (p *parser) parseLabel() (t.ID, error) {
 	return 0, nil
 }
 
+func (p *parser) parseEndwhile(label t.ID) bool {
+	if p.peek1() != t.IDEndwhile {
+		return false
+	}
+	p.src = p.src[1:]
+	if p.peek1() != t.IDDot {
+		return false
+	}
+	p.src = p.src[1:]
+	if p.peek1() != label {
+		return false
+	}
+	p.src = p.src[1:]
+	return true
+}
+
 func (p *parser) parseStatement1() (*a.Node, error) {
 	x := p.peek1()
 	if x == t.IDVar {
@@ -720,6 +736,10 @@ func (p *parser) parseStatement1() (*a.Node, error) {
 		body, err := p.parseBlock()
 		if err != nil {
 			return nil, err
+		}
+		if (label != 0) && !p.parseEndwhile(label) {
+			return nil, fmt.Errorf(`parse: expected endwhile.%s at %s:%d`,
+				label.Str(p.tm), p.filename, p.line())
 		}
 		return a.NewWhile(label, condition, asserts, body).AsNode(), nil
 	}